Added some real time compass display ability, moved commands to their own class
