+++ Android/SensorPack/app/src/main/res/layout/content_main.xml
@@ -3,12 +3,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="16dp"
     android:orientation="vertical"
-    android:weightSum="6.15"
+    android:weightSum="6"
     tools:context="utap.navsea.sensorpack.MainActivity"
     tools:showIn="@layout/activity_main">
 
@@ -46,14 +46,6 @@
         android:src="@drawable/compass" />
         <!--compass picture from:
         https://upload.wikimedia.org/wikipedia/commons/8/8e/Compass_Card.png-->
-
-    <Button
-        android:id="@+id/rtbutton_main"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:layout_gravity="center_vertical|center_horizontal"
-        android:text="@string/button_realtime"
-        android:layout_weight="0.15" />
 
 </LinearLayout>
