+++ Android/SensorPack/app/src/main/res/values/strings.xml
@@ -23,4 +23,5 @@
     <string name="menu_stop">Stop</string>
 
     <string name="button_realtime">Graph real time data</string>
+    <string name="button_realtime_compgyro">Display real time data</string>
 </resources>
